requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
google-auth==2.12.0
idna==3.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.5.0
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
